+++ examples/message_broker.rs
@@ -16,7 +16,6 @@ impl SimpleComponent for Header {
     type Init = ();
     type Input = bool;
     type Output = AppMsg;
-    type Widgets = HeaderWidgets;
 
     view! {
         gtk::HeaderBar {
@@ -92,7 +91,6 @@ impl SimpleComponent for Dialog {
     type Init = gtk::Window;
     type Input = DialogMsg;
     type Output = AppMsg;
-    type Widgets = DialogWidgets;
 
     view! {
         dialog = gtk::MessageDialog {
@@ -163,7 +161,6 @@ impl SimpleComponent for App {
     type Init = ();
     type Input = AppMsg;
     type Output = ();
-    type Widgets = AppWidgets;
 
     view! {
         main_window = gtk::ApplicationWindow {