+++ relm4-macros/tests/builder.rs
@@ -2,10 +2,10 @@ use gtk::prelude::GtkWindowExt;
 use relm4::{gtk, ComponentParts, ComponentSender, SimpleComponent};
 
 #[derive(Default)]
-struct AppModel;
+struct App;
 
 #[relm4_macros::component]
-impl SimpleComponent for AppModel {
+impl SimpleComponent for App {
     type Init = ();
     type Input = ();
     type Output = ();