+++ relm4-macros/tests/simple.rs
@@ -2,7 +2,7 @@ use gtk::prelude::{BoxExt, ButtonExt, GtkWindowExt, OrientableExt};
 use relm4::{gtk, ComponentParts, ComponentSender, RelmWidgetExt, SimpleComponent};
 
 #[derive(Default)]
-struct AppModel {
+struct App {
     counter: u8,
 }
 
@@ -13,11 +13,10 @@ enum AppMsg {
 }
 
 #[relm4_macros::component]
-impl SimpleComponent for AppModel {
+impl SimpleComponent for App {
     type Init = u8;
     type Input = AppMsg;
     type Output = ();
-    type Widgets = AppWidgets;
 
     view! {
         gtk::Window {
@@ -68,11 +67,4 @@ impl SimpleComponent for AppModel {
             }
         }
     }
-}
-
-fn assert_impls_debug<T: std::fmt::Debug>() {}
-
-#[test]
-fn assert_widgets_impl_debug() {
-    assert_impls_debug::<AppWidgets>();
 }