+++ relm4-macros/tests/widget_template.rs
@@ -18,7 +18,7 @@ impl WidgetTemplate for CustomBox {
 }
 
 #[derive(Default)]
-struct AppModel {
+struct App {
     counter: u8,
 }
 
@@ -29,11 +29,10 @@ enum AppMsg {
 }
 
 #[relm4_macros::component]
-impl SimpleComponent for AppModel {
+impl SimpleComponent for App {
     type Init = u8;
     type Input = AppMsg;
     type Output = ();
-    type Widgets = AppWidgets;
 
     view! {
         gtk::Window {
@@ -87,11 +86,4 @@ impl SimpleComponent for AppModel {
             }
         }
     }
-}
-
-fn assert_impls_debug<T: std::fmt::Debug>() {}
-
-#[test]
-fn assert_widgets_impl_debug() {
-    assert_impls_debug::<AppWidgets>();
 }